refactor: 메뉴바 이름 네비게이션바로 변경
